k4niz/AnalytiQ/requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
google-generativeai
python-dotenv
